Harry branch

+++ server.js
@@ -59,8 +59,12 @@ app.post('/fetchComplaint',verify, async (req, res) => {
 });
 
 app.post('/deleteComplaint',verify, async (req, res) => {
-
-  complaint = await db.deleteIncident(req.body.complaintId, req.body.user._id);
+  const userID = jwt.verify(req.body.token, process.env.ACCESS_SECRET);
+  if (userID.admin) {
+    complaint = await db.AdminDeleteIncident(req.body.complaintId);
+  } else {
+    complaint = await db.deleteIncident(req.body.complaintId, req.body.user._id);
+  }
   if(complaint.deletedCount == 1) res.send({status: "report deleted", return: complaint});
   if(complaint.deletedCount != 1) res.send({status: "report not deleted. Please try again"})
 });
@@ -174,6 +178,28 @@ app.post('/verify', async (req,res) => {
   }
 })
 
+app.get('/allComplaints', async (req,res) => {
+  complaints = await db.allIndcidents();
+  res.send(complaints);
+})
+
+app.get('/allStores', async (req,res) => {
+  stores = await db.getAllStores();
+  console.log(stores);
+  res.send(stores);
+})
+
+app.post('/addStore', async (req,res) => {
+  console.log(req.body.sCentre);
+  try {
+    add = await db.addStore(req.body.storeName, req.body.sCentre);
+    console.log(add);
+    res.send({status: "added store successfully"});
+  } catch (error) {
+    console.log(error);
+    res.send({status: "error adding store"})
+  }
+})
 
 app.listen(port, () => {
   console.log(`Example app listening at http://localhost:${port}`)
